refactor(shared): extract conversation.ts and privileged-sites.ts from types.ts (#1912)
types.ts (732 lines) had become a residual dump rather than a barrel:
src/shared/ already holds ~30 domain modules for conversation DRAFT kinds
(conversation-claims-drafts.ts, conversation-note-body-drafts.ts, ...), yet
every conversation CORE type -- ContextBundle, TurnUsage,
ConversationMessage, ConversationStatus, Conversation -- was still stuck in
types.ts. The tell: PrivilegedSite (per-machine cookie-partition state, owned
on the main side by main/privileged-sites.ts) sat inside the "Conversations"
section despite having nothing to do with conversations.

Two new modules:
- shared/conversation.ts -- the missing module for the conversation core:
  ContextBundleNode, ContextBundle, Citation, TurnUsage, ConversationMessage,
  ConversationStatus, CompactResult, ConversationsUIState, ConversationSkill,
  ConversationCreateOptions, Conversation.
- shared/privileged-sites.ts -- just PrivilegedSite, moved to its own
  domain-named module matching main/privileged-sites.ts.

No re-export shim: every one of the ~20 import sites across main, renderer,
shared, and tests now imports directly from the new module, matching how
every other shared/conversation-*.ts domain file is already consumed rather
than adding a third way to reach the same type.

channels.ts and ipc-contract.ts are untouched -- per the issue, those
genuinely are catalogs (one entry per IPC channel), not dumps.

types.ts: 732 -> 559 lines. BUDGETS entry lowered in the same diff; client.ts
grew by one line (an import split in two) and its entry was raised to match.
No behavior change -- pure type relocation. pnpm test
tests/preload/preload-bridge.test.ts confirms the window.api surface is
byte-for-byte unchanged (this only moves where types are imported FROM, not
the runtime bridge shape).
