+++ tests/architecture/file-size-budgets.test.ts
@@ -58,7 +58,7 @@ const BUDGETS: Record<string, number> = {
   'src/renderer/lib/components/Preview.svelte': 1299,
   'src/renderer/lib/components/SourceDetail.svelte': 1337,
   'src/renderer/lib/components/SourcesPanel.svelte': 1292,
-  'src/renderer/lib/ipc/client.ts': 1239,
+  'src/renderer/lib/ipc/client.ts': 1240,
   'src/renderer/lib/stores/conversations.svelte.ts': 1178,
   'src/renderer/lib/components/Editor.svelte': 824,
   'src/renderer/lib/stores/editor.svelte.ts': 1183,
@@ -72,7 +72,7 @@ const BUDGETS: Record<string, number> = {
   'src/renderer/lib/components/ExportDialog.svelte': 712,
   'src/renderer/lib/components/QueryPanel.svelte': 755,
   'src/renderer/lib/components/conversations/DraftCards.svelte': 740,
-  'src/shared/types.ts': 732,
+  'src/shared/types.ts': 559,
   'src/shared/ipc-contract.ts': 720,
   'src/shared/channels.ts': 701,
   'src/renderer/lib/app/note-ops.ts': 687,
